choose self-configuration computer

+++ src/com/Pineapple/Dao/model/Component.java
@@ -4,7 +4,7 @@ public class Component {
 	private String id;// 型号
     private String name;// 名称
     private String type;// 类型
-    private float price;// 价格 
+    private double price;// 价格 
     
   //获取id方法
     public String getId() {
@@ -34,11 +34,11 @@ public class Component {
     }
     
   //获取price方法
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
     //设置price方法
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
